+++ server/index.js
@@ -11,6 +11,11 @@ const router = require('./api/routes.js')
 const User = require('../database/models/user');
 const Destination = require('../database/models/destination');
 const mongoose = require('mongoose')
+// const CronJob = require('cron').CronJob;
+const Cron = require('./cron.js')
+
+
+
 
 // Config variables
 const G_ID = process.env.G_ID || require('./config').G_ID;
